v1.2.1

+++ database/migrations/200_04_08_195233_cidades_migration.php
@@ -14,9 +14,9 @@ class CidadesMigration extends Migration
     public function up()
     {
        Schema::create('cidades', function (Blueprint $table) {
-            $table->increments('id');
+        $table->bigIncrements('id');
             $table->string('descricao_cidade');
-            $table->integer('cod_estado')->unsigned();
+            $table->bigInteger('cod_estado')->unsigned();
             $table->foreign('cod_estado')->references('id')->on('estados');
             $table->timestamps();
         });
